Condense code, added equipment to all horses
After taking time to learn basic Java, I have cleaned up the code significantly and I have made it more efficient. Furthermore, I have added a long-requested feature of spawning in horses with chests, which lets you interact with them IF you are the owner of the horse.

So, for example, if you spawn in a tamed skeleton horse with chests, you will now be able to ride it.